# Improve global bone pose calculation in `Skeleton3D`