@Nazarwadim 2024-09-30
Tested on [Animation_test.zip](https://github.com/godotengine/godot/files/15501000/Animation_test.zip).
- Master: 70 fps
- This PR: 75 fps

As for thread safe (as I understand it you are talking about using the `thread_local` keyword) it is multi-thread safe. Although I do not think that this can cause bugs, since the multi-threaded scene tree is not implemented.