@Saul2022 2024-10-03
> As far as I know multi-threaded scene tree is implemented.

Right, the implementation is here tho, it would be cool in look in this for  skeleton, anim player/tree multithreading improvements.  https://github.com/godotengine/godot/pull/75901