@fire 2024-10-23
@detomon Do you think this can be done to the general 3d node caching system too?