@detomon 2024-10-25
@fire I'm not too familiar with how the Node3D hierarchy works. I will look into that.