@fire 2024-10-25
https://github.com/godotengine/godot/blob/master/scene/3d/node_3d.cpp#L38